Sarah & Abdullah

+++ restaurants/views.py
@@ -42,6 +42,13 @@ def signout(request):
     logout(request)
     return redirect("signin")
 
+
+def no_access(request):
+    context = {
+        "message":"NO WAY!"
+    }
+    return render(request, 'noaccess.html', context)
+
 def restaurant_list(request):
     context = {
         "restaurants":Restaurant.objects.all()
@@ -59,6 +66,8 @@ def restaurant_detail(request, restaurant_id):
     return render(request, 'detail.html', context)
 
 def restaurant_create(request):
+    if request.user.is_anonymous:
+        return redirect('signin')
     form = RestaurantForm()
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES)
@@ -75,6 +84,16 @@ def restaurant_create(request):
 def item_create(request, restaurant_id):
     form = ItemForm()
     restaurant = Restaurant.objects.get(id=restaurant_id)
+    if request.user.is_anonymous:
+        return redirect('signin')
+
+    restaurant_obj = Restaurant.objects.get(id=restaurant_id)
+
+
+    if not (request.user.is_staff or request.user == restaurant_obj.owner):
+
+        return redirect('no-access')
+
     if request.method == "POST":
         form = ItemForm(request.POST)
         if form.is_valid():
@@ -89,7 +108,15 @@ def item_create(request, restaurant_id):
     return render(request, 'item_create.html', context)
 
 def restaurant_update(request, restaurant_id):
+    if request.user.is_anonymous:
+        return redirect('signin')
+
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
+
+
+    if not (request.user.is_staff or request.user == restaurant_obj.owner):
+        return redirect('no-access')
+
     form = RestaurantForm(instance=restaurant_obj)
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES, instance=restaurant_obj)
@@ -104,5 +131,7 @@ def restaurant_update(request, restaurant_id):
 
 def restaurant_delete(request, restaurant_id):
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
+    if not request.user.is_staff:
+        return redirect('no-access')
     restaurant_obj.delete()
     return redirect('restaurant-list')
